Remove turbulence + misc fixes
