+++ Editor/Blocks/Solver.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 0e0b29d957bee114ab6365f3abd45df5
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
