+++ Editor/Blocks/Solver/InitializeSolver.cs
