+++ Editor/Blocks/Solver/SpacePartitioning.cs
