+++ Editor/Utils/FluvioFXAttribute.cs
@@ -13,8 +13,6 @@ namespace Thinksquirrel.FluvioFX.Editor
     {
         public static VFXAttribute DensityPressure => new VFXAttribute("densityPressure", VFXValue.Constant<Vector4>());
         public static VFXAttribute Normal => new VFXAttribute("normal", VFXValue.Constant<Vector4>());
-        public static VFXAttribute VorticityTurbulence =>
-            new VFXAttribute("vorticityTurbulence", VFXValue.Constant<Vector4>());
         public static VFXAttribute Force => new VFXAttribute("force", VFXValue.Constant<Vector3>());
         public static VFXAttribute GridIndex => new VFXAttribute("gridIndex", VFXValue.Constant<uint>());
         public static VFXAttribute NeighborCount => new VFXAttribute("neighborCount", VFXValue.Constant<uint>());
