+++ Editor/Utils/FluvioSettings.cs
@@ -2,5 +2,4 @@ public static class FluvioFXSettings
 {
     public const float kEpsilon = 9.99999944e-11f;
     public const float kMaxSqrVelocityChange = 100.0f;
-    public const float kTurbulenceConstant = 10.0f;
 }
